refactor: centralize config utilities
